requirements.txt
boto.utils
argparse